backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
tzdata==2025.2
transformers==4.51.3
pillow==11.2.1
gTTS==2.5.4
torch==2.7.0+cpu
torchvision==0.22.0+cpu
torchaudio==2.7.0+cpu
sentencepiece==0.2.0